Implement Replay Framework Backend APIs (#2260)
This PR:
1. Moved the `setupReplay` function to the base class of worker and
controller to reduce code duplication.
2. If the log storage URI is defined for an execution, this URI will be
persisted into MySQL as the log location of an execution record.
3. Did a little rework on the execution result cleanup.
4. Now when an execution cleans its result, the log will also be
cleaned.

---------

+++ core/amber/src/main/scala/edu/uci/ics/texera/web/TexeraWebApplication.scala
@@ -5,10 +5,12 @@ import com.fasterxml.jackson.module.scala.DefaultScalaModule
 import com.github.dirkraft.dropwizard.fileassets.FileAssetsBundle
 import com.github.toastshaman.dropwizard.auth.jwt.JwtAuthFilter
 import edu.uci.ics.amber.engine.architecture.controller.{ControllerConfig, Workflow}
+import edu.uci.ics.amber.engine.architecture.logreplay.storage.URILogStorage
 import edu.uci.ics.amber.engine.common.{AmberConfig, AmberUtils}
 import edu.uci.ics.amber.engine.common.client.AmberClient
+import edu.uci.ics.amber.engine.common.virtualidentity.ExecutionIdentity
 import edu.uci.ics.texera.Utils
-import edu.uci.ics.texera.Utils.objectMapper
+import edu.uci.ics.texera.Utils.{maptoStatusCode, objectMapper}
 import edu.uci.ics.texera.web.TexeraWebApplication.scheduleRecurringCallThroughActorSystem
 import edu.uci.ics.texera.web.auth.JwtAuth.jwtConsumer
 import edu.uci.ics.texera.web.auth.{
@@ -17,6 +19,7 @@ import edu.uci.ics.texera.web.auth.{
   UserAuthenticator,
   UserRoleAuthorizer
 }
+import edu.uci.ics.texera.web.model.jooq.generated.tables.pojos.WorkflowExecutions
 import edu.uci.ics.texera.web.resource.auth.{AuthResource, GoogleAuthResource}
 import edu.uci.ics.texera.web.resource._
 import edu.uci.ics.texera.web.resource.dashboard.DashboardResource
@@ -33,7 +36,6 @@ import edu.uci.ics.texera.web.resource.dashboard.user.project.{
 }
 import edu.uci.ics.texera.web.resource.dashboard.user.quota.UserQuotaResource
 import edu.uci.ics.texera.web.resource.dashboard.user.discussion.UserDiscussionResource
-import edu.uci.ics.texera.web.resource.dashboard.user.workflow.WorkflowExecutionsResource.ExecutionResultEntry
 import edu.uci.ics.texera.web.resource.dashboard.user.workflow.{
   WorkflowAccessResource,
   WorkflowExecutionsResource,
@@ -42,7 +44,7 @@ import edu.uci.ics.texera.web.resource.dashboard.user.workflow.{
 }
 import edu.uci.ics.texera.web.service.ExecutionsMetadataPersistService
 import edu.uci.ics.texera.web.storage.MongoDatabaseManager
-import edu.uci.ics.texera.web.workflowruntimestate.WorkflowAggregatedState
+import edu.uci.ics.texera.web.workflowruntimestate.WorkflowAggregatedState.{COMPLETED, FAILED}
 import io.dropwizard.auth.{AuthDynamicFeature, AuthValueFactoryProvider}
 import io.dropwizard.setup.{Bootstrap, Environment}
 import io.dropwizard.websockets.WebsocketBundle
@@ -57,6 +59,7 @@ import scala.concurrent.ExecutionContext.Implicits.global
 import scala.concurrent.duration.{DurationInt, FiniteDuration}
 import org.apache.commons.jcs3.access.exception.InvalidArgumentException
 
+import java.net.URI
 import scala.annotation.tailrec
 
 object TexeraWebApplication {
@@ -128,14 +131,22 @@ class TexeraWebApplication extends io.dropwizard.Application[TexeraWebConfigurat
     bootstrap.addBundle(new WebsocketBundle(classOf[CollaborationResource]))
     // register scala module to dropwizard default object mapper
     bootstrap.getObjectMapper.registerModule(DefaultScalaModule)
-
-    if (AmberConfig.sinkStorageMode.equalsIgnoreCase("mongodb")) {
+    if (AmberConfig.isUserSystemEnabled) {
       val timeToLive: Int = AmberConfig.sinkStorageTTLInSecs
       // do one time cleanup of collections that were not closed gracefully before restart/crash
-      // retrieve all executions that are not completed
-      val incompleteExecutions: List[ExecutionResultEntry] =
-        WorkflowExecutionsResource.getAllIncompleteResults()
-      cleanOldCollections(incompleteExecutions, WorkflowAggregatedState.FAILED)
+      // retrieve all executions that were executing before the reboot.
+      val allExecutionsBeforeRestart: List[WorkflowExecutions] =
+        WorkflowExecutionsResource.getExpiredExecutionsWithResultOrLog(-1)
+      cleanExecutions(
+        allExecutionsBeforeRestart,
+        statusByte => {
+          if (statusByte != maptoStatusCode(COMPLETED)) {
+            maptoStatusCode(FAILED) // for incomplete executions, mark them as failed.
+          } else {
+            statusByte
+          }
+        }
+      )
       scheduleRecurringCallThroughActorSystem(
         2.seconds,
         AmberConfig.sinkStorageCleanUpCheckIntervalInSecs.seconds
@@ -225,27 +236,42 @@ class TexeraWebApplication extends io.dropwizard.Application[TexeraWebConfigurat
     * MongoDB doesn't have an API of drop collection where collection name in (from a subquery), so the implementation is to retrieve
     * the entire list of those documents that have expired, then loop the list to drop them one by one
     */
-  def cleanOldCollections(
-      executions: List[ExecutionResultEntry],
-      status: WorkflowAggregatedState
+  private def cleanExecutions(
+      executions: List[WorkflowExecutions],
+      statusChangeFunc: Byte => Byte
   ): Unit = {
     // drop the collection and update the status to ABORTED
     executions.foreach(execEntry => {
-      dropCollections(execEntry.result)
+      dropCollections(execEntry.getResult)
+      deleteReplayLog(execEntry.getLogLocation)
       // then delete the pointer from mySQL
-      ExecutionsMetadataPersistService.tryUpdateExecutionStatusAndPointers(
-        execEntry.eId.longValue(),
-        status
-      )
+      val executionIdentity = ExecutionIdentity(execEntry.getEid.longValue())
+      ExecutionsMetadataPersistService.tryUpdateExistingExecution(executionIdentity) { execution =>
+        execution.setResult("")
+        execution.setLogLocation(null)
+        execution.setStatus(statusChangeFunc(execution.getStatus))
+      }
     })
   }
 
   def dropCollections(result: String): Unit = {
+    if (result.isEmpty) {
+      return
+    }
     // parse the JSON
     val node = objectMapper.readTree(result)
     val collections = node.get("results")
     // loop every collection and drop it
     collections.forEach(collection => MongoDatabaseManager.dropCollection(collection.asText()))
+  }
+
+  def deleteReplayLog(logLocation: String): Unit = {
+    if (logLocation == null || logLocation.isEmpty) {
+      return
+    }
+    val uri = new URI(logLocation)
+    val storage = new URILogStorage(uri)
+    storage.deleteStorage()
   }
 
   /**
@@ -255,9 +281,9 @@ class TexeraWebApplication extends io.dropwizard.Application[TexeraWebConfigurat
       timeToLive: Int
   ): Unit = {
     // retrieve all executions that are completed and their last update time goes beyond the ttl
-    val expiredResults: List[ExecutionResultEntry] =
-      WorkflowExecutionsResource.getExpiredResults(timeToLive)
-    // drop the collection and update the status to COMPLETED
-    cleanOldCollections(expiredResults, WorkflowAggregatedState.COMPLETED)
+    val expiredResults: List[WorkflowExecutions] =
+      WorkflowExecutionsResource.getExpiredExecutionsWithResultOrLog(timeToLive)
+    // drop the collections and clean the logs
+    cleanExecutions(expiredResults, statusByte => statusByte)
   }
 }
